requirements.txt
Django==1.11.3
djangorestframework==3.6.3
django-rest-swagger==2.1.2
coreapi==2.3.1
